+++ src/components/controllPanel/presentation/fragments/tabs/ControlPanelTabController.tsx
@@ -1,6 +1,6 @@
 import { faShare } from "@fortawesome/free-solid-svg-icons";
 import React, { FunctionComponent } from "react";
-import { ModuleRoutes } from "../../../../../app/Davit";
+import { ModuleRoutes } from "../../../../../pages/Davit";
 import { DavitButton } from "../../../../common/fragments/buttons/DavitButton";
 import { ControlPanel } from "../edit/common/ControlPanel";
 import { OptionField } from "../edit/common/OptionField";
