+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import { Davit } from "./app/Davit";
+import { Davit } from "./pages/Davit";
 import { store } from "./app/store";
 import * as serviceWorker from "./serviceWorker";
 import "./style/index.css";
