+++ src/pages/Davit.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from "@testing-library/react";
 import React from "react";
 import { Provider } from "react-redux";
+import { store } from "../app/store";
 import { Davit } from "./Davit";
-import { store } from "./store";
 
 test("Davit basic UI test", () => {
     render(
